Updating to newest wasmcloud SDK API
