+++ host_core/test/e2e/kvcounter_test.exs
@@ -125,6 +125,11 @@ defmodule HostCore.E2E.KVCounterTest do
     {:ok, resp} = request_http("http://localhost:8081/foobar", 2)
     assert resp.body == "{\"counter\":#{incr_count}}"
 
+    {:ok, body} = JSON.decode(resp.body)
+    incr_count = Map.get(body, "counter", 0) + 1
+    {:ok, resp} = request_http("http://localhost:8081/foobar", 2)
+    assert resp.body == "{\"counter\":#{incr_count}}"
+
     # :timer.sleep(6000)
   end
